+++ elements/reigns/src/features/voting/calculateAnswer.test.ts
@@ -1,5 +1,5 @@
 import { getSdk } from "../../sdk";
-import { mockSdk } from "../game/mocks";
+import { mockSdk } from "../../mocks";
 import { calculateAnswer } from "./calculateAnswer";
 import { persistParticipantVote } from "./persistence";
 import { PARTICIPANT_INSIDE_TABLE } from "./useOnFrescoStateUpdate";
