+++ elements/reigns/src/features/voting/triggerEventOnParticipantVote.test.ts
@@ -1,6 +1,6 @@
 import { Countdown } from "../../Countdown";
 import { getSdk } from "../../sdk";
-import { mockSdk } from "../game/mocks";
+import { mockSdk } from "../../mocks";
 import { Answer, persistParticipantVote } from "./persistence";
 import { triggerEventOnParticipantVote } from "./triggerEventOnParticipantVote";
 import { PARTICIPANT_INSIDE_TABLE } from "./useOnFrescoStateUpdate";
